backend/requirements.txt
tensorflow
fastapi
uvicorn
python-multipart
pillow
numpy
scipy
streamlit
requests
python-dotenv
pandas
streamlit-option-menu
